@@ -1,9 +1,10 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+// import ReactDOM from 'react-dom';
+import ReactDOM from "react-dom/client";
+
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
-import { useTheme, createTheme, ThemeProvider } from '@mui/material/styles';
+import { createTheme, ThemeProvider } from '@mui/material/styles';
 import { Provider } from 'react-redux';
 import { store } from './Redux/Store';
 
@@ -12,14 +13,14 @@ const theme = createTheme({
     fontFamily: ['Kanit', 'sans-serif'].join(','),
   },
 });
-ReactDOM.render(
-  <React.StrictMode>
+const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
+root.render(
+<React.StrictMode>
     <ThemeProvider theme={theme}>
       <Provider store={store}>
         <App />
       </Provider>
     </ThemeProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
-);
-reportWebVitals();
+  </React.StrictMode>
+)
+
